+++ models/microsoft/graph/action_stepable.go
@@ -9,9 +9,9 @@ type ActionStepable interface {
     i04eb5309aeaafadd28374d79c8471df9b267510b4dc2e3144c378c50f6fd7b55.AdditionalDataHolder
     i04eb5309aeaafadd28374d79c8471df9b267510b4dc2e3144c378c50f6fd7b55.Parsable
     GetActionUrl()(ActionUrlable)
-    GetStepNumber()(*int64)
+    GetStepNumber()(*int32)
     GetText()(*string)
     SetActionUrl(value ActionUrlable)()
-    SetStepNumber(value *int64)()
+    SetStepNumber(value *int32)()
     SetText(value *string)()
 }
