+++ models/microsoft/graph/certificate_connector_health_metric_valueable.go
@@ -10,9 +10,9 @@ type CertificateConnectorHealthMetricValueable interface {
     i04eb5309aeaafadd28374d79c8471df9b267510b4dc2e3144c378c50f6fd7b55.AdditionalDataHolder
     i04eb5309aeaafadd28374d79c8471df9b267510b4dc2e3144c378c50f6fd7b55.Parsable
     GetDateTime()(*i336074805fc853987abe6f7fe3ad97a6a6f3077a16391fec744f671a015fbd7e.Time)
-    GetFailureCount()(*int64)
-    GetSuccessCount()(*int64)
+    GetFailureCount()(*int32)
+    GetSuccessCount()(*int32)
     SetDateTime(value *i336074805fc853987abe6f7fe3ad97a6a6f3077a16391fec744f671a015fbd7e.Time)()
-    SetFailureCount(value *int64)()
-    SetSuccessCount(value *int64)()
+    SetFailureCount(value *int32)()
+    SetSuccessCount(value *int32)()
 }
