+++ models/microsoft/graph/device_log_collection_response.go
@@ -9,7 +9,7 @@ import (
 type DeviceLogCollectionResponse struct {
     Entity
     // The error code, if any. Valid values -9.22337203685478E+18 to 9.22337203685478E+18
-    errorCode *int64;
+    errorCode *int32;
     // The DateTime of the expiration of the logs
     expirationDateTimeUTC *i336074805fc853987abe6f7fe3ad97a6a6f3077a16391fec744f671a015fbd7e.Time;
     // The UPN for who initiated the request
@@ -37,7 +37,7 @@ func CreateDeviceLogCollectionResponseFromDiscriminatorValue(parseNode i04eb5309
     return NewDeviceLogCollectionResponse(), nil
 }
 // GetErrorCode gets the errorCode property value. The error code, if any. Valid values -9.22337203685478E+18 to 9.22337203685478E+18
-func (m *DeviceLogCollectionResponse) GetErrorCode()(*int64) {
+func (m *DeviceLogCollectionResponse) GetErrorCode()(*int32) {
     if m == nil {
         return nil
     } else {
@@ -56,7 +56,7 @@ func (m *DeviceLogCollectionResponse) GetExpirationDateTimeUTC()(*i336074805fc85
 func (m *DeviceLogCollectionResponse) GetFieldDeserializers()(map[string]func(interface{}, i04eb5309aeaafadd28374d79c8471df9b267510b4dc2e3144c378c50f6fd7b55.ParseNode)(error)) {
     res := m.Entity.GetFieldDeserializers()
     res["errorCode"] = func (o interface{}, n i04eb5309aeaafadd28374d79c8471df9b267510b4dc2e3144c378c50f6fd7b55.ParseNode) error {
-        val, err := n.GetInt64Value()
+        val, err := n.GetInt32Value()
         if err != nil {
             return err
         }
@@ -195,7 +195,7 @@ func (m *DeviceLogCollectionResponse) Serialize(writer i04eb5309aeaafadd28374d79
         return err
     }
     {
-        err = writer.WriteInt64Value("errorCode", m.GetErrorCode())
+        err = writer.WriteInt32Value("errorCode", m.GetErrorCode())
         if err != nil {
             return err
         }
@@ -245,7 +245,7 @@ func (m *DeviceLogCollectionResponse) Serialize(writer i04eb5309aeaafadd28374d79
     return nil
 }
 // SetErrorCode sets the errorCode property value. The error code, if any. Valid values -9.22337203685478E+18 to 9.22337203685478E+18
-func (m *DeviceLogCollectionResponse) SetErrorCode(value *int64)() {
+func (m *DeviceLogCollectionResponse) SetErrorCode(value *int32)() {
     if m != nil {
         m.errorCode = value
     }
