+++ models/microsoft/graph/device_log_collection_responseable.go
@@ -9,7 +9,7 @@ import (
 type DeviceLogCollectionResponseable interface {
     Entityable
     i04eb5309aeaafadd28374d79c8471df9b267510b4dc2e3144c378c50f6fd7b55.Parsable
-    GetErrorCode()(*int64)
+    GetErrorCode()(*int32)
     GetExpirationDateTimeUTC()(*i336074805fc853987abe6f7fe3ad97a6a6f3077a16391fec744f671a015fbd7e.Time)
     GetInitiatedByUserPrincipalName()(*string)
     GetManagedDeviceId()(*string)
@@ -17,7 +17,7 @@ type DeviceLogCollectionResponseable interface {
     GetRequestedDateTimeUTC()(*i336074805fc853987abe6f7fe3ad97a6a6f3077a16391fec744f671a015fbd7e.Time)
     GetSize()(*float64)
     GetStatus()(*string)
-    SetErrorCode(value *int64)()
+    SetErrorCode(value *int32)()
     SetExpirationDateTimeUTC(value *i336074805fc853987abe6f7fe3ad97a6a6f3077a16391fec744f671a015fbd7e.Time)()
     SetInitiatedByUserPrincipalName(value *string)()
     SetManagedDeviceId(value *string)()
