+++ models/microsoft/graph/education_synchronization_profile_statusable.go
@@ -9,12 +9,12 @@ import (
 type EducationSynchronizationProfileStatusable interface {
     Entityable
     i04eb5309aeaafadd28374d79c8471df9b267510b4dc2e3144c378c50f6fd7b55.Parsable
-    GetErrorCount()(*int64)
+    GetErrorCount()(*int32)
     GetLastActivityDateTime()(*i336074805fc853987abe6f7fe3ad97a6a6f3077a16391fec744f671a015fbd7e.Time)
     GetLastSynchronizationDateTime()(*i336074805fc853987abe6f7fe3ad97a6a6f3077a16391fec744f671a015fbd7e.Time)
     GetStatus()(*EducationSynchronizationStatus)
     GetStatusMessage()(*string)
-    SetErrorCount(value *int64)()
+    SetErrorCount(value *int32)()
     SetLastActivityDateTime(value *i336074805fc853987abe6f7fe3ad97a6a6f3077a16391fec744f671a015fbd7e.Time)()
     SetLastSynchronizationDateTime(value *i336074805fc853987abe6f7fe3ad97a6a6f3077a16391fec744f671a015fbd7e.Time)()
     SetStatus(value *EducationSynchronizationStatus)()
