+++ models/microsoft/graph/key_long_value_pairable.go
@@ -9,7 +9,7 @@ type KeyLongValuePairable interface {
     i04eb5309aeaafadd28374d79c8471df9b267510b4dc2e3144c378c50f6fd7b55.AdditionalDataHolder
     i04eb5309aeaafadd28374d79c8471df9b267510b4dc2e3144c378c50f6fd7b55.Parsable
     GetName()(*string)
-    GetValue()(*int64)
+    GetValue()(*int32)
     SetName(value *string)()
-    SetValue(value *int64)()
+    SetValue(value *int32)()
 }
